+++ src/storage/internals/storage_internal.rs
@@ -1,4 +1,3 @@
-use js_sys::Object;
 use wasm_bindgen::JsValue;
 use wasm_bindgen::prelude::wasm_bindgen;
 use crate::operation::Operation;
@@ -63,6 +62,7 @@ export abstract class StorageInternal<T extends SchemaType> {
      */
     abstract close(): Promise<void>;
 }
+
 /**
  * Represents a function type for creating storage with the provided schema type records.
  *
@@ -73,7 +73,6 @@ export abstract class StorageInternal<T extends SchemaType> {
 export type CreateStorage = <T extends SchemaTypeRecord = SchemaTypeRecord>(
     records: T
 ) => InternalsRecord;
-
 "#;
 
 #[wasm_bindgen]
@@ -81,14 +80,14 @@ extern "C" {
     #[derive(Clone, Default)]
     pub type StorageInternal;
 
-    #[wasm_bindgen(constructor)]
-    pub fn new(name: &JsValue, schema: &JsValue) -> StorageInternal;
-
     #[wasm_bindgen(method, getter = schema)]
     pub fn schema(this: &StorageInternal) -> Schema;
 
     #[wasm_bindgen(method, getter = name)]
     pub fn name(this: &StorageInternal) -> String;
+
+    #[wasm_bindgen(method, getter = migration)]
+    pub fn migration(this: &StorageInternal) -> JsValue;
 
     #[wasm_bindgen(method, catch)]
     pub async fn write(this: &StorageInternal, op: Operation) -> Result<JsValue, JsValue>;